% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMethods.R
\name{plot.prediction_errors}
\alias{plot.prediction_errors}
\title{Function to plot comparison of Prediction methods}
\usage{
\method{plot}{prediction_errors}(x, ...)
}
\arguments{
\item{x}{as output object of 'prediction_errors()' function}

\item{...}{arguments passed to or from other methods}
}
\value{
Returns error comparison plots for forecasting methods
}
\description{
Function to plot comparison of Prediction methods
}
\examples{
a <- prediction_errors(data = nottem)
b <- plot(a)
}
