% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_circle.R
\name{plot_circle}
\alias{plot_circle}
\title{Function to plot comparison of Predicted values in a circular ring}
\usage{
plot_circle(x, ...)
}
\arguments{
\item{x}{as output object of 'prediction_errors()' function}

\item{...}{arguments passed to or from other methods}
}
\value{
Returns error comparison plots for forecasting methods
}
\description{
Function to plot comparison of Predicted values in a circular ring
}
\examples{
a <- prediction_errors(data = nottem)
plot_circle(a)
}
