% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_errors.R
\name{prediction_errors}
\alias{prediction_errors}
\title{Function working as testbench for comparison of Prediction methods}
\usage{
prediction_errors(
  data,
  nval,
  ePara,
  ePara_name,
  Method,
  MethodName,
  strats,
  dval,
  append_
)
}
\arguments{
\item{data}{as input time series for testing}

\item{nval}{as an integer to decide number of values to predict}

\item{ePara}{as type of error calculation (RMSE and MAE are default), add an error parameter of
your choice in the following manner: ePara = c("errorparametername")
where errorparametername is should be a source/function which returns desired error set}

\item{ePara_name}{as list of  names of error parameters passed in order}

\item{Method}{as the list of locations of function for the proposed prediction method
(should be recursive) (default:arima)}

\item{MethodName}{as list of names for function for the proposed prediction method in order}

\item{strats}{as list of forecasting strategies. Available : recursive and dirrec}

\item{dval}{as last d values of the data to be used for forecasting}

\item{append_}{suggests if the function is used to append to another instance}
}
\value{
Returns error comparison for forecasting methods
}
\description{
Function working as testbench for comparison of Prediction methods
}
\examples{
prediction_errors(data = nottem)
}
